+++ temporal_gradient/clock/chronos.py
@@ -1,7 +1,6 @@
 import math
 import textwrap
 import time
-import warnings
 
 from temporal_gradient.salience.pipeline import KeywordImperativeValue, RollingJaccardNovelty, SaliencePipeline
 from temporal_gradient.telemetry.chronometric_vector import ChronometricVector
@@ -38,24 +37,6 @@ class ClockRateModulator:
         self.legacy_density_scale = legacy_density_scale
         self.strict_psi_bounds = strict_psi_bounds
         self.chronology = []
-
-    @property
-    def chronolog(self):
-        warnings.warn(
-            "ClockRateModulator.chronolog is deprecated; use chronology instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self.chronology
-
-    @chronolog.setter
-    def chronolog(self, value):
-        warnings.warn(
-            "ClockRateModulator.chronolog is deprecated; use chronology instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        self.chronology = value
 
     def _validate_salience_mode(self, salience_mode):
         valid_modes = {"canonical", "legacy_density"}